# Optimize detect_duplicate_ids on parse_document and parse_fragment