@ypconstante 2025-09-02
@SteffenDE moved to `Phoenix.LiveViewTest.DOM` and removed all changes that I had made to `Phoenix.LiveViewTest.TreeDOM`. New `detect_duplicate_ids` is a private function, so no need to check the struct type.